This is a quick PR in response to @jchiang87's [Photometry Bakeoff](https://github.com/jchiang87/PhotometryBakeoff) results.  This was using the GalSim Bandpass and SED classes in a way that we never really tried to optimize for speed.  So it turned out the code wasn't terribly fast.  

Here (on branch table_speedup) I made a few simple changes to speed things up.

1. Use pandas to read in the text file if it's available, falling back to the old (and much slower) `numpy.loadtxt` function.  This wasn't actually part of Jim's timings, since he started the clock after the SED and Bandpass were both loaded.  But it makes a huge difference in the overall running time of the galsim_baker script.

2. Add a function to the C++ Table class that can interpolate many values at once.  This saves a huge overhead from the python generator that was doing the interpolation one item at a time and `numpy.fromiter` to turn the values into an array.

3. Use `numpy.min` and `numpy.max` rather than the python `min`, `max` for arrays in a couple places.  This is a standard optimization.  I guess we just failed to notice these before.

4. Use the `Table.argMin()` and `argMax()` functions for the min and max table values rather than recalculating them for the `x_min` and `x_max` properties.  This is another silly oversight, since the values are already known, so there is no point to calculating them again.

Results:

The total time for running Jim's galsim_baker.py script on my laptop with and without thinning the bandpass:

Current master branch:

Without Thinning: 50m 10s
With Thinning: 29m 59s

This branch:

Without Thinning: 3m 47s
With Thinning: 2m 54s

So an order of magnitude speed up.  Not bad for a day's work... :)