@rmandelb 2016-03-30
Mike - what version of pandas do you have?  I'm getting loads of test failures because I do have pandas, so it tries to use it, but my version of pandas (0.6.1) does not have a ```comment``` kwarg for ```read_csv```, so whenever it hits the call to ```read_csv```, it barfs.

I can of course update pandas on my system, but we might want to do a version check before trying to use pandas to ensure that we're on a version that can handle this syntax.