@rmandelb 2016-03-30
I upgraded pandas to the latest release.  Now I only get one test error:

```
ERROR: Basic tests of SED functionality
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/Library/Frameworks/EPD64.framework/Versions/7.2/lib/python2.7/site-packages/nose/case.py", line 197, in runTest
    self.test(*self.arg)
  File "/Users/rmandelb/git/GalSim/tests/test_sed.py", line 100, in test_SED_basic
    np.testing.assert_almost_equal(s(400)*h*c/400, 200, decimal=10)
  File "/Library/Frameworks/EPD64.framework/Versions/7.2/lib/python2.7/site-packages/galsim/sed.py", line 213, in __call__
    .format(wmax, self.red_limit))
ValueError: Requested wavelength (400) is redder than red_limit (6.90775527898)
-------------------- >> begin captured stdout << ---------------------
0  s =  galsim.SED(<function <lambda> at 0x1125047d0>, redshift=0.0)
1  s =  galsim.SED('200', redshift=0.0)
Try pickling  galsim.SED('200', redshift=0.0)
Try pickling  galsim.SED('200', redshift=0.0)
2  s =  galsim.SED('200', redshift=0.0)
Try pickling  galsim.SED('200', redshift=0.0)
Try pickling  galsim.SED('200', redshift=0.0)
3  s =  galsim.SED('200', redshift=0.0)
Try pickling  galsim.SED('200', redshift=0.0)
Try pickling  galsim.SED('200', redshift=0.0)
4  s =  galsim.SED('20', redshift=0.0)
Try pickling  galsim.SED('20', redshift=0.0)
Try pickling  galsim.SED('20', redshift=0.0)
5  s =  galsim.SED('200 * wave / 1.986445683269303e-09', redshift=0.0)
Try pickling  galsim.SED('200 * wave / 1.986445683269303e-09', redshift=0.0)
Try pickling  galsim.SED('200 * wave / 1.986445683269303e-09', redshift=0.0)
6  s =  galsim.SED('20 * (wave/10) / 1.986445683269303e-09', redshift=0.0)
Try pickling  galsim.SED('20 * (wave/10) / 1.986445683269303e-09', redshift=0.0)
Try pickling  galsim.SED('20 * (wave/10) / 1.986445683269303e-09', redshift=0.0)
7  s =  galsim.SED('200 * wave**2 / 2.99792458e+17', redshift=0.0)
Try pickling  galsim.SED('200 * wave**2 / 2.99792458e+17', redshift=0.0)
Try pickling  galsim.SED('200 * wave**2 / 2.99792458e+17', redshift=0.0)
8  s =  galsim.SED('200 * (wave/10)**2 / 2.99792458e+17', redshift=0.0)
Try pickling  galsim.SED('200 * (wave/10)**2 / 2.99792458e+17', redshift=0.0)
Try pickling  galsim.SED('200 * (wave/10)**2 / 2.99792458e+17', redshift=0.0)
9  s =  galsim.SED(galsim.LookupTable(x=[1.0,..,1000.0], f=[200.0,...,200.0], interpolant='linear'), redshift=0.0)
Try pickling  galsim.SED(galsim.LookupTable(x=[1.0,..,1000.0], f=[200.0,...,200.0], interpolant='linear'), redshift=0.0)
Try pickling  galsim.SED(galsim.LookupTable(x=[1.0,..,1000.0], f=[200.0,...,200.0], interpolant='linear'), redshift=0.0)
10  s =  galsim.SED(galsim.LookupTable(x=[1.0,..,10000.0], f=[20.0,...,20.0], interpolant='linear'), redshift=0.0)
Try pickling  galsim.SED(galsim.LookupTable(x=[1.0,..,10000.0], f=[20.0,...,20.0], interpolant='linear'), redshift=0.0)
Try pickling  galsim.SED(galsim.LookupTable(x=[1.0,..,10000.0], f=[20.0,...,20.0], interpolant='linear'), redshift=0.0)
11  s =  galsim.SED(galsim.LookupTable(x=[1.0,..,1000.0], f=[100682340164.0,...,1.00682340164e+14], interpolant='linear'), redshift=0.0)
Try pickling  galsim.SED(galsim.LookupTable(x=[1.0,..,1000.0], f=[100682340164.0,...,1.00682340164e+14], interpolant='linear'), redshift=0.0)
Try pickling  galsim.SED(galsim.LookupTable(x=[1.0,..,1000.0], f=[100682340164.0,...,1.00682340164e+14], interpolant='linear'), redshift=0.0)
12  s =  galsim.SED(galsim.LookupTable(x=[1.0,..,10000.0], f=[1006823401.64,...,1.00682340164e+13], interpolant='linear'), redshift=0.0)
Try pickling  galsim.SED(galsim.LookupTable(x=[1.0,..,10000.0], f=[1006823401.64,...,1.00682340164e+13], interpolant='linear'), redshift=0.0)
Try pickling  galsim.SED(galsim.LookupTable(x=[1.0,..,10000.0], f=[1006823401.64,...,1.00682340164e+13], interpolant='linear'), redshift=0.0)
13  s =  galsim.SED(galsim.LookupTable(x=[1.0,..,1000.0], f=[6.67128190396e-16,...,6.67128190396e-10], interpolant='linear'), redshift=0.0)

--------------------- >> end captured stdout << ----------------------
```