@rmandelb 2016-03-30
Also, I just noticed I get this failure on my cluster too, where I don't have pandas at all.