@rmjarvis 2016-03-30
I have 0.16.2.  Probably best to add an AttributeError to the `except` statement to catch that too.

Also, I have that same unit test failure.  Sorry.  I forgot to run scons tests on the final version.  I'll look into it.