@rmjarvis 2016-03-30
> ValueError: Requested wavelength (400) is redder than red_limit (6.90775527898)

Fixed.