@rmandelb 2016-04-04
I pushed a commit ( https://github.com/GalSim-developers/GalSim/pull/735/commits/08838a291faf1d343a7476e844a5868f0e8b5296 ) that changed a few min/max in chromatic.py to np.min/np.max, similar to what you did here in bandpass.py and sed.py.

I think it would be worth putting pandas in the list of "optional dependencies" in INSTALL.md.  Also this stuff is important enough to go in CHANGELOG.  The code looks good to me and tests pass on both of my systems.