@rmjarvis 2016-04-04
Thanks Rachel.  Done on both counts.